clean up room viewer routes
* refactor urls to be nicer
* clean up room model code to include url property for viewer
* dry out the code
